Rename component/presentational to component/base
