+++ src/component/base/BarMeter.js
