+++ src/component/base/Slider.js
